chore: move session tests to AmplitudeSessionTests
